Phi: fixed forgotten cast of statistical weight from unsigned int to double
This caused a lot of LTE level populations to be zero or NaN due to
decimal truncation.

+++ src/EOS/Phi.cc
@@ -10,13 +10,13 @@
 double Phi(const AtomicLevel &level, const Ion &ion, const Atom &atom, const double &temperature) {
     const double C_I = 0.5 * std::pow(std::pow(h_planck, 2) / (2.0 * pi * m_electron * k_boltzmann), 3.0/2.0); // Mihalas defines this constant in his Eq (5-14)
 
-    return (level.g / ion.continuum_state->g) * C_I * std::pow(temperature, -3.0/2.0) * std::exp((ion.ionization_potential - level.energy) / (k_boltzmann * temperature));
+    return (double(level.g) / double(ion.continuum_state->g)) * C_I * std::pow(temperature, -3.0/2.0) * std::exp((ion.ionization_potential - level.energy) / (k_boltzmann * temperature));
 }
 
 
 // Mihalas defines the Phi_tilde function in his Eq (5-15)
 double Phi_tilde(const AtomicLevel &level, const Ion &ion, const Atom &atom, const double &temperature) {
-    return (ion.continuum_state->g / ion.partition_function) * Phi(level, ion, atom, temperature);
+    return (double(ion.continuum_state->g) / ion.partition_function) * Phi(level, ion, atom, temperature);
 }
 
 // Mihalas defines the other Phi_tilde function in his Eq (5-16)
